Road to Playtesting- Update 1
The first of a series of 10-12 updates that will prepare the mod for
playtesting.

>3 New Bookmarks (Not different times, just different locations, like
the different 1444 bookmarks in base EU4)
>AI will now actually build some of the available buildings
>8 new tags added to southern Oaxaca. I finally found a reliable map of
the Mixtec Kingdom of Tututepec and the towns around and in it.
>Borders of many provinces that were really pissing me off redrawn
>Borders of several provinces redrawn so that you could get historical
Tarascan and Aztec borders
>Ugly, choppy coastline fixed
>Coastline redrawn in several areas for increased realism
>Advisors balanced for the mod
>Slight adjustments to a few techs
>Provinces will now get negative modifiers the more times they are
occupied within a short span of time, and can lose population if
sufficiently damaged.
>Triggered modifiers for different levels of manpower added.
>New vassal liberty desire system implemented pretty much fully - you
slowly build up reduced liberty desire with your vassals based on a
number of factors. You can quickly lose this by losing battles and wars
or becoming unstable.
>"Assess the Loyalty of our Subjects" decision added
>Minor bugfixes