The Road to Playtesting - Update 2
Second of a series of updates designed to prepare the mod for
playtesting.

Changes:

>Several "Meta" events (population changes, LPP setup) are no longer
visible to the player!
>Finally fixed that bug where a country would get 20-30 new monarchs in
the span of a second
>Added government stats
>Changed some buildings to boost government stats
>Added administrative influence - local autonomy for vassals
>Occupying a province for long enough may result in the inhabitants of
the city surrendering to you - they immediately become your tributary or
territory without you making a peace deal
>Units cost more to recruit and less to maintain, and reinforce slower
>Minor balance changes and bugfixes
>Custom main theme added TEMPORARILY - I can't use it because it's
copyrighted, but it sounds nice and this version isn't being
distributed. Will be removed before release.

>IMPORTANT: There are two major bugs that may still exist. The first is
vassal liberty desire SKYROCKETING for a reason I haven't figured out.
The second is you becoming the tributary of another country and that
same country becoming your tributary. I think the second one is fixed,
but I'm not 100% sure.